Cleanup hardware extensions/capabilities
